requirements.txt
asgiref==3.7.2
Django==4.2.4
django-froala-editor==4.1.1
greenlet==2.0.2
Pillow==10.0.0
playwright==1.37.0
pyee==9.0.4
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
